Refactoring to remove unused base classes to move to MVP.
