--- app/src/main/java/com/robotandpencils/app/ui/common/BaseChildFragmentModule.java
@@ -1,15 +1,0 @@
-package com.robotandpencils.app.ui.common;
-
-import dagger.Module;
-
-/**
- * Created by pwray on 2017-09-27.
- *
- * The class `BaseChildFragmentModule` provides the base child fragment dependencies.
- */
-
-@Module
-public abstract class BaseChildFragmentModule {
-
-    public static final String CHILD_FRAGMENT = "BaseChildFragmentModule.childFragment";
-}
