+++ app/src/main/java/com/robotandpencils/app/ui/example_3/parent_fragment/view/Example3ParentFragmentModule.java
@@ -4,7 +4,7 @@ import android.app.Fragment;
 
 import com.robotandpencils.app.inject.PerChildFragment;
 import com.robotandpencils.app.inject.PerFragment;
-import com.robotandpencils.app.ui.common.BaseFragmentModule;
+import com.robotandpencils.app.ui.common.view.BaseFragmentModule;
 import com.robotandpencils.app.ui.example_3.child_fragment.view.Example3ChildFragment;
 import com.robotandpencils.app.ui.example_3.child_fragment.view.Example3ChildFragmentModule;
 import com.robotandpencils.app.ui.example_3.parent_fragment.presenter.Example3ParentPresenterModule;
